Added method to create a client with admin rights
This method creates a client with the 'admin' flag set, in an open source
Chef Server.

This does not work in Hosted Chef, since Opscode Platform
security design uses a different approach. By default a client is only
allowed to manage the node it's associated with, and setting this flag
in the POST or PUT request has no effect on the client permissions.