Provide the existing validator credentials in context creation
This commit fixes #12.

Now when bootstrapping nodes using Chef, the validator credentials
must be provided in context creation. Since we cannot create validator
clients, having those credentials is a requirement to make the nodes
be able to autoregister to the Chef server during bootstrap.
When using this feature, the following properties must be set:
    ChefProperties.CHEF_VALIDATOR_NAME
    ChefProperties.CHEF_VALIDATOR_CREDENTIAL

Also refactored ChefConstants class into ChefProperties to follow the
new conventions.